+++ GenericModConfigMenu/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* The mod list now remembers your scroll position when returning from a config view (thanks to Pathoschild!).
+
 ## 1.8.1
 Released 12 January 2022 for SMAPI 3.13.0 or later. Updated by Pathoschild.